Resolve conflicts by moving files to biome_configuration
