+++ crates/biome_configuration/src/diagnostics.rs
@@ -1,17 +1,15 @@
+use crate::WorkspaceError;
 use biome_console::fmt::Display;
 use biome_console::{markup, MarkupBuf};
 use biome_deserialize::DeserializationDiagnostic;
 use biome_diagnostics::adapters::ResolveError;
-use biome_diagnostics::{
-    Advices, Category, Diagnostic, DiagnosticTags, Error, Location, LogCategory,
-    MessageAndDescription, Severity, Visit,
-};
+use biome_diagnostics::{Advices, Diagnostic, Error, LogCategory, MessageAndDescription, Visit};
 use biome_rowan::SyntaxError;
 use serde::{Deserialize, Serialize};
 use std::fmt::{Debug, Formatter};
 
 /// Series of errors that can be thrown while computing the configuration
-#[derive(serde::Serialize, serde::Deserialize)]
+#[derive(Deserialize, Diagnostic, Serialize)]
 pub enum ConfigurationDiagnostic {
     /// Thrown when the program can't serialize the configuration, while saving it
     SerializationError(SerializationError),
@@ -53,11 +51,11 @@ impl From<DeserializationDiagnostic> for ConfigurationDiagnostic {
 }
 
 impl ConfigurationDiagnostic {
-    pub fn new_serialization_error() -> Self {
+    pub(crate) fn new_serialization_error() -> Self {
         Self::SerializationError(SerializationError)
     }
 
-    pub fn new_invalid_ignore_pattern(
+    pub(crate) fn new_invalid_ignore_pattern(
         pattern: impl Into<String>,
         reason: impl Into<String>,
     ) -> Self {
@@ -105,116 +103,6 @@ impl std::fmt::Display for ConfigurationDiagnostic {
     }
 }
 
-impl Diagnostic for ConfigurationDiagnostic {
-    fn severity(&self) -> Severity {
-        match self {
-            ConfigurationDiagnostic::SerializationError(error) => error.severity(),
-            ConfigurationDiagnostic::ConfigAlreadyExists(error) => error.severity(),
-            ConfigurationDiagnostic::Deserialization(error) => error.severity(),
-            ConfigurationDiagnostic::InvalidIgnorePattern(error) => error.severity(),
-            ConfigurationDiagnostic::CantLoadExtendFile(error) => error.severity(),
-            ConfigurationDiagnostic::InvalidConfiguration(error) => error.severity(),
-            ConfigurationDiagnostic::CantResolve(error) => error.severity(),
-        }
-    }
-
-    fn category(&self) -> Option<&'static Category> {
-        match self {
-            ConfigurationDiagnostic::SerializationError(error) => error.category(),
-            ConfigurationDiagnostic::ConfigAlreadyExists(error) => error.category(),
-            ConfigurationDiagnostic::Deserialization(error) => error.category(),
-            ConfigurationDiagnostic::InvalidIgnorePattern(error) => error.category(),
-            ConfigurationDiagnostic::CantLoadExtendFile(error) => error.category(),
-            ConfigurationDiagnostic::InvalidConfiguration(error) => error.category(),
-            ConfigurationDiagnostic::CantResolve(error) => error.category(),
-        }
-    }
-
-    fn tags(&self) -> DiagnosticTags {
-        match self {
-            ConfigurationDiagnostic::SerializationError(error) => error.tags(),
-            ConfigurationDiagnostic::ConfigAlreadyExists(error) => error.tags(),
-            ConfigurationDiagnostic::Deserialization(error) => error.tags(),
-            ConfigurationDiagnostic::InvalidIgnorePattern(error) => error.tags(),
-            ConfigurationDiagnostic::CantLoadExtendFile(error) => error.tags(),
-            ConfigurationDiagnostic::InvalidConfiguration(error) => error.tags(),
-            ConfigurationDiagnostic::CantResolve(error) => error.tags(),
-        }
-    }
-
-    fn location(&self) -> Location<'_> {
-        match self {
-            ConfigurationDiagnostic::SerializationError(error) => error.location(),
-            ConfigurationDiagnostic::ConfigAlreadyExists(error) => error.location(),
-            ConfigurationDiagnostic::Deserialization(error) => error.location(),
-            ConfigurationDiagnostic::InvalidIgnorePattern(error) => error.location(),
-            ConfigurationDiagnostic::CantLoadExtendFile(error) => error.location(),
-            ConfigurationDiagnostic::InvalidConfiguration(error) => error.location(),
-            ConfigurationDiagnostic::CantResolve(error) => error.location(),
-        }
-    }
-
-    fn source(&self) -> Option<&dyn Diagnostic> {
-        match self {
-            ConfigurationDiagnostic::SerializationError(error) => error.source(),
-            ConfigurationDiagnostic::ConfigAlreadyExists(error) => error.source(),
-            ConfigurationDiagnostic::Deserialization(error) => error.source(),
-            ConfigurationDiagnostic::InvalidIgnorePattern(error) => error.source(),
-            ConfigurationDiagnostic::CantLoadExtendFile(error) => error.source(),
-            ConfigurationDiagnostic::InvalidConfiguration(error) => error.source(),
-            ConfigurationDiagnostic::CantResolve(error) => error.source(),
-        }
-    }
-
-    fn message(&self, fmt: &mut biome_console::fmt::Formatter<'_>) -> std::io::Result<()> {
-        match self {
-            ConfigurationDiagnostic::SerializationError(error) => error.message(fmt),
-            ConfigurationDiagnostic::ConfigAlreadyExists(error) => error.message(fmt),
-            ConfigurationDiagnostic::Deserialization(error) => error.message(fmt),
-            ConfigurationDiagnostic::InvalidIgnorePattern(error) => error.message(fmt),
-            ConfigurationDiagnostic::CantLoadExtendFile(error) => error.message(fmt),
-            ConfigurationDiagnostic::InvalidConfiguration(error) => error.message(fmt),
-            ConfigurationDiagnostic::CantResolve(error) => error.message(fmt),
-        }
-    }
-
-    fn description(&self, fmt: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ConfigurationDiagnostic::SerializationError(error) => error.description(fmt),
-            ConfigurationDiagnostic::ConfigAlreadyExists(error) => error.description(fmt),
-            ConfigurationDiagnostic::Deserialization(error) => error.description(fmt),
-            ConfigurationDiagnostic::InvalidIgnorePattern(error) => error.description(fmt),
-            ConfigurationDiagnostic::CantLoadExtendFile(error) => error.description(fmt),
-            ConfigurationDiagnostic::InvalidConfiguration(error) => error.description(fmt),
-            ConfigurationDiagnostic::CantResolve(error) => error.description(fmt),
-        }
-    }
-
-    fn advices(&self, visitor: &mut dyn Visit) -> std::io::Result<()> {
-        match self {
-            ConfigurationDiagnostic::SerializationError(error) => error.advices(visitor),
-            ConfigurationDiagnostic::ConfigAlreadyExists(error) => error.advices(visitor),
-            ConfigurationDiagnostic::Deserialization(error) => error.advices(visitor),
-            ConfigurationDiagnostic::InvalidIgnorePattern(error) => error.advices(visitor),
-            ConfigurationDiagnostic::CantLoadExtendFile(error) => error.advices(visitor),
-            ConfigurationDiagnostic::InvalidConfiguration(error) => error.advices(visitor),
-            ConfigurationDiagnostic::CantResolve(error) => error.advices(visitor),
-        }
-    }
-
-    fn verbose_advices(&self, visitor: &mut dyn Visit) -> std::io::Result<()> {
-        match self {
-            ConfigurationDiagnostic::SerializationError(error) => error.verbose_advices(visitor),
-            ConfigurationDiagnostic::ConfigAlreadyExists(error) => error.verbose_advices(visitor),
-            ConfigurationDiagnostic::Deserialization(error) => error.verbose_advices(visitor),
-            ConfigurationDiagnostic::InvalidIgnorePattern(error) => error.verbose_advices(visitor),
-            ConfigurationDiagnostic::CantLoadExtendFile(error) => error.verbose_advices(visitor),
-            ConfigurationDiagnostic::InvalidConfiguration(error) => error.verbose_advices(visitor),
-            ConfigurationDiagnostic::CantResolve(error) => error.verbose_advices(visitor),
-        }
-    }
-}
-
 #[derive(Debug, Serialize, Deserialize, Default)]
 pub struct ConfigurationAdvices {
     messages: Vec<MarkupBuf>,
@@ -254,7 +142,7 @@ pub struct ConfigAlreadyExists {}
 pub struct InvalidIgnorePattern {
     #[message]
     #[description]
-    pub message: String,
+    pub(crate) message: String,
 }
 
 #[derive(Debug, Serialize, Deserialize, Diagnostic)]
@@ -290,6 +178,12 @@ impl CantLoadExtendFile {
     }
 }
 
+impl From<CantLoadExtendFile> for WorkspaceError {
+    fn from(value: CantLoadExtendFile) -> Self {
+        WorkspaceError::Configuration(ConfigurationDiagnostic::CantLoadExtendFile(value))
+    }
+}
+
 #[derive(Debug, Serialize, Deserialize, Diagnostic)]
 #[diagnostic(
 	category = "configuration",
@@ -318,8 +212,8 @@ pub struct CantResolve {
 
 #[cfg(test)]
 mod test {
-    use crate::diagnostics::ConfigurationDiagnostic;
-    use crate::PartialConfiguration;
+    use crate::configuration::diagnostics::ConfigurationDiagnostic;
+    use crate::configuration::PartialConfiguration;
     use biome_deserialize::json::deserialize_from_json_str;
     use biome_diagnostics::{print_diagnostic_to_string, DiagnosticExt, Error};
     use biome_json_parser::JsonParserOptions;
